Create RSIbot.py
The Relative Strength Index (RSI) is calculated for a specific trading pair "symbol" (BTC/USDT). Then, it checks the RSI conditions and executes buy or sell operations based on the specified conditions (an infinite loop calculates the RSI and performs trading operations according to the logic defined in the trade function).

When the RSI drops below 20, a flag (buyFlag) is set, indicating a potential buying opportunity. If the RSI drops below 10 and there is available balance in the quote currency (BUSD), another flag (buyFlag2) is set. Buys are executed when the RSI exceeds 21, and an additional buy is triggered if the RSI surpasses 10. On the other hand, a flag (sellFlag) is set to indicate a potential selling opportunity when the RSI exceeds 65. Sells are executed if the RSI falls below 65 or exceeds 75. These conditions form a basic trading system aiming to capitalize on overbought and oversold conditions according to the RSI.

You only have to change variables apiKey and secret for the ones you created on Binance API gestor.
Also, there are some variables you can change, depending your method: par1 and par2 to combine the symbol you want to trade (the one used here is BTC/USDT); interval and period of the RSI calculated (I calculate the RSI of the last 8 periods of 1 minute).